+++ src/JoinRpg.Interfaces/Email/EmailSendFailedException.cs
@@ -1,0 +1,5 @@
+namespace JoinRpg.Interfaces.Email;
+
+public class EmailSendFailedException(string message) : Exception(message)
+{
+}
