+++ src/JoinRpg.Interfaces/Email/IEmailSendingService.cs
@@ -1,6 +1,6 @@
 using JoinRpg.DataModel;
 
-namespace JoinRpg.Services.Interfaces.Email;
+namespace JoinRpg.Interfaces.Email;
 
 public interface IEmailSendingService
 {
@@ -15,6 +15,7 @@ public interface IEmailSendingService
         RecepientData recepient)
      => SendEmails(subject, body, sender, [recepient,]);
 
+    [Obsolete("Send markdown emails")]
     Task SendEmails(string subject,
         string body,
         string text,
