+++ src/JoinRpg.Services.Email/AccountServiceEmailImpl.cs
@@ -1,6 +1,5 @@
 using JoinRpg.DataModel;
-using JoinRpg.Services.Interfaces;
-using JoinRpg.Services.Interfaces.Email;
+using JoinRpg.Interfaces.Email;
 using JoinRpg.Services.Interfaces.Notification;
 
 namespace JoinRpg.Services.Email;
