--- src/JoinRpg.Services.Interfaces/Email/EmailSendFailedException.cs
@@ -1,8 +1,0 @@
-namespace JoinRpg.Services.Interfaces.Email;
-
-public class EmailSendFailedException : Exception
-{
-    public EmailSendFailedException(string message) : base(message)
-    {
-    }
-}
