+++ src/JoinRpg.Services.Interfaces/JoinRpg.Services.Interfaces.csproj
@@ -11,7 +11,6 @@
     <ProjectReference Include="..\PscbApi\PscbApi.csproj" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="JetBrains.Annotations" Version="2022.3.1" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
   </ItemGroup>
 </Project>
